#14 - Split Twig_Environment from TwigRendererFactory
This requires the `Zend\Expressive\Twig\TwigEnvironmentFactory::class` to be registered as a factory in the config.
